Add FCN32 results

+++ augment.py
@@ -21,7 +21,7 @@ def resize_with_pad(image, label, size=512):
         (array<np.uint8>): Resized label. Shape=(size, size, 1)
     '''
     image = tf.image.resize_with_pad(image, size, size, method=tf.image.ResizeMethod.NEAREST_NEIGHBOR)
-    ## since default pads with zeros, use fact that boundary class is -1 to pad with -1 instead.
+    ## since `resize_with_pad` pads with zeros, use fact that boundary class is -1 to pad with -1 instead.
     label = tf.image.resize_with_pad(label+1, size, size, method=tf.image.ResizeMethod.NEAREST_NEIGHBOR)-1
-    label = utils.label_to_onehot(label)
+    label = tf.cast(utils.label_to_onehot(label), tf.float32)
     return image, label
